requirements.txt
fastapi
uvicorn
pymongo
voyageai
pydantic
python-dotenv
requests
langchain
langchain-openai
reportlab
